refactor: move to common tests
